feat: refactor Text component and update import path to common directory
