+++ src/components/common/text.tsx
